requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
requests==2.32.3
pydantic==2.11.3
pydantic-settings==2.8.1
python-jose==3.4.0